+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Audit logs in promtail config.
+- Add condition for PSP installation in helm chart.
 
 ## [0.0.6] - 2023-09-28