+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Audit logs in promtail config.
 - Add condition for PSP installation in helm chart.
 
+### Changed
+
+- Logs labels updated to ease navigation.
+
 ## [0.0.6] - 2023-09-28
 
 ### Fixed